+++ .gitignore
@@ -7,4 +7,3 @@ google*.html
 [Tt]humbs.db
 Desktop.ini
 $RECYCLE.BIN/
-composer.lock
