+++ composer.json
@@ -19,5 +19,13 @@
         "phroute/phroute": "^2.1",
         "illuminate/database": "5.2.*",
         "doctrine/dbal": "~2.4"
+    },
+    "autoload": {
+        "psr-4": {
+            "Sakura\\": "libraries/"
+        },
+        "files": [
+            "utility.php"
+        ]
     }
 }
