+++ templates/yuuno/global/master.twig
@@ -50,7 +50,6 @@
 
                 "siteName":         "{{ config('sitename') }}",
                 "content":          "{{ config('content_path') }}",
-                "resources":        "{{ sakura.resources }}",
                 "recaptchaEnabled": "{{ config('recaptcha') }}",
 
                 "minUserLen":       {{ config('username_min_length') }},
