--- templates/yuuno/old-settings/general.home.twig
@@ -1,20 +1,0 @@
-<div style="margin: 5px;">
-    <h1 class="stylised">Common Tasks</h1>
-    <h2>Profile</h2>
-    <ul>
-        <li><a href="{{ urls.format('SETTING_MODE', ['appearance', 'avatar']) }}" class="default">Change Avatar</a></li>
-        <li><a href="{{ urls.format('SETTING_MODE', ['appearance', 'userpage']) }}" class="default">Change Userpage</a></li>
-        <li><a href="{{ urls.format('SETTING_MODE', ['appearance', 'signature']) }}" class="default">Change Signature</a></li>
-        <li><a href="{{ urls.format('SETTING_MODE', ['general', 'profile']) }}" class="default">Change Profile Details</a></li>
-    </ul>
-    <h2>Messaging</h2>
-    <ul>
-        <li><a href="{{ urls.format('SETTING_MODE', ['messages', 'inbox']) }}" class="default">View Inbox</a></li>
-        <li><a href="{{ urls.format('SETTING_MODE', ['messages', 'compose']) }}" class="default">Send PM</a></li>
-    </ul>
-    <h2>Account</h2>
-    <ul>
-        <li><a href="{{ urls.format('SETTING_MODE', ['advanced', 'sessions']) }}" class="default">Manage Active Sessions</a></li>
-        <li><a href="{{ urls.format('SETTING_MODE', ['account', 'password']) }}" class="default">Change Password</a></li>
-    </ul>
-</div>
