+++ templates/yuuno/settings/advanced/sessions.twig
@@ -1,0 +1,51 @@
+{% extends 'settings/advanced/master.twig' %}
+
+{% set mode = 'Home' %}
+
+{% block description %}
+    <p>Welcome to the Settings Panel! From here you can monitor, view and update your profile and preferences.</p>
+{% endblock %}
+
+{% block settingsContent %}
+    <table class="settings-table">
+        <thead>
+            <tr><th style="width: 100px;">IP</th><th>Useragent</th><th style="width: 120px;">Login time</th><th></th></tr>
+        </thead>
+        <tfoot>
+            <tr><th>IP</th><th>Useragent</th><th>Login time</th><th></th></tr>
+        </tfoot>
+        <tbody>
+            {% for s in sessions %}
+            <tr {% if s.session_key == session.sessionId %} class="current-session"{% endif %}>
+                <td>
+                    {{ s.user_ip }}
+                </td>
+                <td>
+                    {{ s.user_agent }}
+                </td>
+                <td>
+                    <time datetime="{{ s.session_start|date('r') }}">{{ s.session_start|date(config('date_format')) }}</time>
+                </td>
+                <td style="width: 90px;">
+                    <form method="post" action="/settings.php?cat=advanced&amp;mode=sessions">
+                        <input type="hidden" name="sessid" value="{{ session_id() }}" />
+                        <input type="hidden" name="timestamp" value="{{ date().timestamp }}" />
+                        <input type="hidden" name="sessionid" value="{{ s.session_id }}" />
+                        <input type="hidden" name="mode" value="sessions" />
+                        <button class="inputStyling small" name="submit">Kill</button>
+                    </form>
+                </td>
+            </tr>
+            {% endfor %}
+        </tbody>
+    </table>
+    <div class="profile-save">
+        <form method="post" action="/settings.php?cat=advanced&amp;mode=sessions">
+            <input type="hidden" name="sessid" value="{{ session_id() }}" />
+            <input type="hidden" name="timestamp" value="{{ date().timestamp }}" />
+            <input type="hidden" name="sessionid" value="all" />
+            <input type="hidden" name="mode" value="sessions" />
+            <button class="inputStyling" name="submit">Kill all active sessions</button>
+        </form>
+    </div>
+{% endblock %}
