+++ templates/yuuno/settings/friends/master.twig
@@ -1,0 +1,11 @@
+{% extends 'settings/master.twig' %}
+
+{% set category = 'Friends' %}
+
+{% block css %}
+    <style type="text/css">
+        .pagination {
+            float: right;
+        }
+    </style>
+{% endblock %}
