+++ templates/yuuno/settings/notifications/master.twig
@@ -1,0 +1,3 @@
+{% extends 'settings/master.twig' %}
+
+{% set category = 'Notifications' %}
