+++ templates/yuuno/user/report.twig
@@ -1,0 +1,5 @@
+{% extends 'global/master.twig' %}
+
+{% block content %}
+    <h1 class="stylised" style="text-align: center; margin: 2em auto;">I'll actually make reporting a thing, someday...</h1>
+{% endblock %}
